refactor many defs of math utils to be in lib/utils with very simple plugin code
